Move page objects from webPages to pages
