+++ scenarioo-client/test/e2e/pages/editorPage.ts
