+++ scenarioo-client/test/e2e/pages/labelConfigurationsPage.ts
