+++ scenarioo-client/test/e2e/pages/navigationPage.ts
