+++ scenarioo-client/test/e2e/pages/objectDetailsPage.ts
