+++ scenarioo-client/test/e2e/specs/browse_page_variants.ts
@@ -2,8 +2,8 @@
 
 import { scenario, step, useCase } from 'scenarioo-js';
 import * as Utils from '../util/util';
-import StepPage from '../webPages/stepPage';
-import ObjectDetailsPage from '../webPages/objectDetailsPage';
+import StepPage from '../pages/stepPage';
+import ObjectDetailsPage from '../pages/objectDetailsPage';
 
 useCase('Browse page variants')
     .description('On the step page the user can flip through all existing variants of the page.')
