+++ scenarioo-client/test/e2e/specs/configure_label_colors.ts
@@ -2,8 +2,8 @@
 
 import { scenario, step, useCase } from 'scenarioo-js';
 import * as Utils from '../util/util';
-import HomePage from '../webPages/homePage';
-import LabelConfigurationsPage from '../webPages/labelConfigurationsPage';
+import HomePage from '../pages/homePage';
+import LabelConfigurationsPage from '../pages/labelConfigurationsPage';
 
 useCase('Configure label colors')
     .description('Each label string can be configured to be displayed in a certain color.')
