+++ scenarioo-client/test/e2e/specs/full_text_search.ts
@@ -2,12 +2,12 @@
 
 import { scenario, step, useCase } from 'scenarioo-js';
 import * as Utils from '../util/util';
-import HomePage from '../webPages/homePage';
-import StepPage from '../webPages/stepPage';
-import BreadcrumbsPage from '../webPages/breadcrumbsPage';
-import SearchResultsPage from '../webPages/searchResultsPage';
-import NavigationPage from '../webPages/navigationPage';
-import GeneralSettingsPage from '../webPages/generalSettingsPage';
+import HomePage from '../pages/homePage';
+import StepPage from '../pages/stepPage';
+import BreadcrumbsPage from '../pages/breadcrumbsPage';
+import SearchResultsPage from '../pages/searchResultsPage';
+import NavigationPage from '../pages/navigationPage';
+import GeneralSettingsPage from '../pages/generalSettingsPage';
 
 useCase('Full Text Search')
     .description('Search within all information of the selected build.')
