+++ scenarioo-client/test/e2e/specs/list_custom_tabs.ts
@@ -2,7 +2,7 @@
 
 import { scenario, step, useCase } from 'scenarioo-js';
 import * as Utils from '../util/util';
-import HomePage from '../webPages/homePage';
+import HomePage from '../pages/homePage';
 
 useCase('List custom tabs')
     .description('Custom tabs can be defined to show aggregated documentation data..')
