+++ scenarioo-client/test/e2e/specs/list_scenarios.ts
@@ -2,10 +2,10 @@
 
 import { scenario, step, useCase } from 'scenarioo-js';
 import * as Utils from '../util/util';
-import HomePage from '../webPages/homePage';
-import UsecasePage from '../webPages/usecasePage';
-import ScenarioPage from '../webPages/scenarioPage';
-import NavigationPage from '../webPages/navigationPage';
+import HomePage from '../pages/homePage';
+import UsecasePage from '../pages/usecasePage';
+import ScenarioPage from '../pages/scenarioPage';
+import NavigationPage from '../pages/navigationPage';
 
 const NUMBER_OF_USE_CASES = 4;
 const NUMBER_OF_SCENARIOS = 4;
