+++ scenarioo-client/test/e2e/specs/manage_branch_aliases.ts
@@ -2,7 +2,7 @@
 
 import { scenario, step, useCase } from 'scenarioo-js';
 import * as Utils from '../util/util';
-import BranchAliasesPage from '../webPages/branchAliasesPage';
+import BranchAliasesPage from '../pages/branchAliasesPage';
 
 const NUMBER_OF_ALIASES_IN_CONFIG = 2;
 const FIRST_TEST_ALIAS_INDEX = NUMBER_OF_ALIASES_IN_CONFIG;
